+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - 16-02-2021
+### Added
+- Routing cache
+- Internal queue metrics
+
+### Updated
+- Command subscription QoS levels (from 1 to 3)
+- Load tests for new caches
+
+### Removed
+- Ingress projects
+- API projects (except the network API)
+- Common projects (caching)
+
 ## [1.3.0] - 06-02-2021
 ### Updated
 - Measurement query result JSON deserialization